perf/api/scale: use() alloc reuse, Instance throw, NA accessor, crash run_id int64
Section B — the deferred fresh-angle-review findings, each verified against
current code first (3 read-only probes), then only the real ones fixed:

- use() per-call allocations removed (speed-first win, xi_use.hpp). process()
  rebuilt in_imgs/in_handles every frame and exchange() malloc'd a fresh 64KB
  buffer every call; both now reuse per-thread scratch. Safe because use() calls
  never nest on a thread — process_fn/exchange_fn are plugins, and a plugin can't
  call xi::use() (it's script-side), so they can't re-enter with the buffers live.
- Instance<T> deref now throws instead of UB-on-null (xi_instance.hpp). The
  comment promised "first dereference will throw"; the operators just returned
  ptr_.get() and crashed. Added a checked_() helper behind all four ->/* operators
  carrying the instance name; `if (inst)` still tests without throwing.
- Typed::has() + Number::value_opt() (xi_types.hpp) — additive checked reads that
  tell "field present and 0" apart from "missing / NA"; value() keeps its
  silent-downgrade behaviour. Addresses the filed per-field-NA backlog item.
  Unit test added (test_record.cpp 16c).
- Crash breadcrumb last_run_id widened int->int64 (xi_crash_dump.hpp +
  service_main.cpp): it was truncating the real int64 run_id, so a crash after
  2^31 runs reported a wrong/negative id even though the wire run_id is int64.

Deliberately NOT changed (verified non-issues / disproportionate): vars-snapshot
copy (RVO-free in Release; string->buffer is the C-ABI boundary); history ring
copy (bounded 50, post-run, copy required); as_cv_mat const-split (OpenCV has no
const-correctness — already the buglist const-input-ABI item); g_status cap
(refuted — keys are bounded by instance count); the xi_inspect_entry(int) frame
hint wrap (user-facing script-ABI; deferred to the next ABI bump, now filed in
known-issues.md).

Verified: Release all-targets clean; unit 0 fail (incl. new test); ws_comprehensive
/ run_vars / trigger / preview / defect / commit_group green.

+++ backend/include/xi/xi_crash_dump.hpp
@@ -74,8 +74,8 @@ struct Context {
     char last_plugin[64]   {};   // plugin name backing it
     char last_phase[32]    {};   // inspect lifecycle phase (reset/inspect/...)
     char last_status[96]   {};   // last xi::status()/set_status text on this thread
-    int  last_run_id       = 0;
-    int  last_frame        = 0;
+    int64_t last_run_id    = 0;  // full run id (wire run_id is int64; don't truncate)
+    int  last_frame        = 0;  // frame hint — int by the xi_inspect_entry(int) ABI
 };
 
 // Per-thread crash breadcrumbs. A single global was racy under dispatch_threads
